pyproject.toml
[project]
name = "memoryai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "dspy>=3.1.2",
    "numpy>=2.4.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "uuid>=1.30",
]
